+++ service/codebuild/api.go
@@ -4898,6 +4898,12 @@ type Build struct {
 	Environment *ProjectEnvironment `locationName:"environment" type:"structure"`
 
 	// A list of exported environment variables for this build.
+	//
+	// Exported environment variables are used in conjunction with AWS CodePipeline
+	// to export environment variables from the current build stage to subsequent
+	// stages in the pipeline. For more information, see Working with variables
+	// (https://docs.aws.amazon.com/codepipeline/latest/userguide/actions-variables.html)
+	// in the AWS CodePipeline User Guide.
 	ExportedEnvironmentVariables []*ExportedEnvironmentVariable `locationName:"exportedEnvironmentVariables" type:"list"`
 
 	// An array of ProjectFileSystemLocation objects for a CodeBuild build project.
@@ -5213,6 +5219,38 @@ type BuildArtifacts struct {
 	// An identifier for this artifact definition.
 	ArtifactIdentifier *string `locationName:"artifactIdentifier" type:"string"`
 
+	// Specifies the access for objects that are uploaded to an Amazon S3 bucket
+	// that is owned by another account.
+	//
+	// By default, only the account that uploads the objects to the bucket has access
+	// to these objects. This property allows you to give the bucket owner access
+	// to these objects.
+	//
+	// NONE
+	//
+	// The bucket owner does not have access to the objects. This is the default.
+	//
+	// READ_ONLY
+	//
+	// The bucket owner has read only access to the objects. The uploading account
+	// retains ownership of the objects.
+	//
+	// FULL
+	//
+	// The bucket owner has full access to the objects. Object ownership is determined
+	// by the following criteria:
+	//
+	//    * If the bucket is configured with the Bucket owner preferred setting,
+	//    the bucket owner owns the objects. The uploading account will have object
+	//    access as specified by the bucket's policy.
+	//
+	//    * Otherwise, the uploading account retains ownership of the objects.
+	//
+	// For more information about Amazon S3 object ownership, see Controlling ownership
+	// of uploaded objects using S3 Object Ownership (https://docs.aws.amazon.com/AmazonS3/latest/userguide/about-object-ownership.html)
+	// in the Amazon Simple Storage Service User Guide.
+	BucketOwnerAccess *string `locationName:"bucketOwnerAccess" type:"string" enum:"BucketOwnerAccess"`
+
 	// Information that tells you if encryption for build artifacts is disabled.
 	EncryptionDisabled *bool `locationName:"encryptionDisabled" type:"boolean"`
 
@@ -5257,6 +5295,12 @@ func (s BuildArtifacts) GoString() string {
 // SetArtifactIdentifier sets the ArtifactIdentifier field's value.
 func (s *BuildArtifacts) SetArtifactIdentifier(v string) *BuildArtifacts {
 	s.ArtifactIdentifier = &v
+	return s
+}
+
+// SetBucketOwnerAccess sets the BucketOwnerAccess field's value.
+func (s *BuildArtifacts) SetBucketOwnerAccess(v string) *BuildArtifacts {
+	s.BucketOwnerAccess = &v
 	return s
 }
 
@@ -7943,19 +7987,25 @@ func (s *EnvironmentVariable) SetValue(v string) *EnvironmentVariable {
 	return s
 }
 
-// Information about an exported environment variable.
+// Contains information about an exported environment variable.
+//
+// Exported environment variables are used in conjunction with AWS CodePipeline
+// to export environment variables from the current build stage to subsequent
+// stages in the pipeline. For more information, see Working with variables
+// (https://docs.aws.amazon.com/codepipeline/latest/userguide/actions-variables.html)
+// in the AWS CodePipeline User Guide.
+//
+// During a build, the value of a variable is available starting with the install
+// phase. It can be updated between the start of the install phase and the end
+// of the post_build phase. After the post_build phase ends, the value of exported
+// variables cannot change.
 type ExportedEnvironmentVariable struct {
 	_ struct{} `type:"structure"`
 
-	// The name of this exported environment variable.
+	// The name of the exported environment variable.
 	Name *string `locationName:"name" min:"1" type:"string"`
 
-	// The value assigned to this exported environment variable.
-	//
-	// During a build, the value of a variable is available starting with the install
-	// phase. It can be updated between the start of the install phase and the end
-	// of the post_build phase. After the post_build phase ends, the value of exported
-	// variables cannot change.
+	// The value assigned to the exported environment variable.
 	Value *string `locationName:"value" type:"string"`
 }
 
@@ -10214,6 +10264,38 @@ type ProjectArtifacts struct {
 	// An identifier for this artifact definition.
 	ArtifactIdentifier *string `locationName:"artifactIdentifier" type:"string"`
 
+	// Specifies the access for objects that are uploaded to an Amazon S3 bucket
+	// that is owned by another account.
+	//
+	// By default, only the account that uploads the objects to the bucket has access
+	// to these objects. This property allows you to give the bucket owner access
+	// to these objects.
+	//
+	// NONE
+	//
+	// The bucket owner does not have access to the objects. This is the default.
+	//
+	// READ_ONLY
+	//
+	// The bucket owner has read only access to the objects. The uploading account
+	// retains ownership of the objects.
+	//
+	// FULL
+	//
+	// The bucket owner has full access to the objects. Object ownership is determined
+	// by the following criteria:
+	//
+	//    * If the bucket is configured with the Bucket owner preferred setting,
+	//    the bucket owner owns the objects. The uploading account will have object
+	//    access as specified by the bucket's policy.
+	//
+	//    * Otherwise, the uploading account retains ownership of the objects.
+	//
+	// For more information about Amazon S3 object ownership, see Controlling ownership
+	// of uploaded objects using S3 Object Ownership (https://docs.aws.amazon.com/AmazonS3/latest/userguide/about-object-ownership.html)
+	// in the Amazon Simple Storage Service User Guide.
+	BucketOwnerAccess *string `locationName:"bucketOwnerAccess" type:"string" enum:"BucketOwnerAccess"`
+
 	// Set to true if you do not want your output artifacts encrypted. This option
 	// is valid only if your artifacts type is Amazon S3. If this is set with another
 	// artifacts type, an invalidInputException is thrown.
@@ -10353,6 +10435,12 @@ func (s *ProjectArtifacts) Validate() error {
 // SetArtifactIdentifier sets the ArtifactIdentifier field's value.
 func (s *ProjectArtifacts) SetArtifactIdentifier(v string) *ProjectArtifacts {
 	s.ArtifactIdentifier = &v
+	return s
+}
+
+// SetBucketOwnerAccess sets the BucketOwnerAccess field's value.
+func (s *ProjectArtifacts) SetBucketOwnerAccess(v string) *ProjectArtifacts {
+	s.BucketOwnerAccess = &v
 	return s
 }
 
@@ -12074,6 +12162,38 @@ func (s *RetryBuildOutput) SetBuild(v *Build) *RetryBuildOutput {
 type S3LogsConfig struct {
 	_ struct{} `type:"structure"`
 
+	// Specifies the access for objects that are uploaded to an Amazon S3 bucket
+	// that is owned by another account.
+	//
+	// By default, only the account that uploads the objects to the bucket has access
+	// to these objects. This property allows you to give the bucket owner access
+	// to these objects.
+	//
+	// NONE
+	//
+	// The bucket owner does not have access to the objects. This is the default.
+	//
+	// READ_ONLY
+	//
+	// The bucket owner has read only access to the objects. The uploading account
+	// retains ownership of the objects.
+	//
+	// FULL
+	//
+	// The bucket owner has full access to the objects. Object ownership is determined
+	// by the following criteria:
+	//
+	//    * If the bucket is configured with the Bucket owner preferred setting,
+	//    the bucket owner owns the objects. The uploading account will have object
+	//    access as specified by the bucket's policy.
+	//
+	//    * Otherwise, the uploading account retains ownership of the objects.
+	//
+	// For more information about Amazon S3 object ownership, see Controlling ownership
+	// of uploaded objects using S3 Object Ownership (https://docs.aws.amazon.com/AmazonS3/latest/userguide/about-object-ownership.html)
+	// in the Amazon Simple Storage Service User Guide.
+	BucketOwnerAccess *string `locationName:"bucketOwnerAccess" type:"string" enum:"BucketOwnerAccess"`
+
 	// Set to true if you do not want your S3 build log output encrypted. By default
 	// S3 build logs are encrypted.
 	EncryptionDisabled *bool `locationName:"encryptionDisabled" type:"boolean"`
@@ -12114,6 +12234,12 @@ func (s *S3LogsConfig) Validate() error {
 		return invalidParams
 	}
 	return nil
+}
+
+// SetBucketOwnerAccess sets the BucketOwnerAccess field's value.
+func (s *S3LogsConfig) SetBucketOwnerAccess(v string) *S3LogsConfig {
+	s.BucketOwnerAccess = &v
+	return s
 }
 
 // SetEncryptionDisabled sets the EncryptionDisabled field's value.
@@ -14659,6 +14785,56 @@ func AuthType_Values() []string {
 		AuthTypeOauth,
 		AuthTypeBasicAuth,
 		AuthTypePersonalAccessToken,
+	}
+}
+
+// Specifies the access for objects that are uploaded to an Amazon S3 bucket
+// that is owned by another account.
+//
+// By default, only the account that uploads the objects to the bucket has access
+// to these objects. This property allows you to give the bucket owner access
+// to these objects.
+//
+// NONE
+//
+// The bucket owner does not have access to the objects. This is the default.
+//
+// READ_ONLY
+//
+// The bucket owner has read only access to the objects. The uploading account
+// retains ownership of the objects.
+//
+// FULL
+//
+// The bucket owner has full access to the objects. Object ownership is determined
+// by the following criteria:
+//
+//    * If the bucket is configured with the Bucket owner preferred setting,
+//    the bucket owner owns the objects. The uploading account will have object
+//    access as specified by the bucket's policy.
+//
+//    * Otherwise, the uploading account retains ownership of the objects.
+//
+// For more information about Amazon S3 object ownership, see Controlling ownership
+// of uploaded objects using S3 Object Ownership (https://docs.aws.amazon.com/AmazonS3/latest/userguide/about-object-ownership.html)
+// in the Amazon Simple Storage Service User Guide.
+const (
+	// BucketOwnerAccessNone is a BucketOwnerAccess enum value
+	BucketOwnerAccessNone = "NONE"
+
+	// BucketOwnerAccessReadOnly is a BucketOwnerAccess enum value
+	BucketOwnerAccessReadOnly = "READ_ONLY"
+
+	// BucketOwnerAccessFull is a BucketOwnerAccess enum value
+	BucketOwnerAccessFull = "FULL"
+)
+
+// BucketOwnerAccess_Values returns all elements of the BucketOwnerAccess enum
+func BucketOwnerAccess_Values() []string {
+	return []string{
+		BucketOwnerAccessNone,
+		BucketOwnerAccessReadOnly,
+		BucketOwnerAccessFull,
 	}
 }